(PDB-2678) Change default cache dir to 'cache'
This commit changes the default directory from 'proxy-cache' to
'cache'. It also simplifies the retrieval of environment variables and
removes the dependency on environ. There was a possibility of errors
causing temp files to be left on disk. This commit adds a try/finally
block to ensure those temp files get cleaned up in the event of a failure